SEguimos con el error, intentar dar de alta un nuevo tipo

+++ Dominio/Importadas.cs
@@ -6,9 +6,9 @@ namespace Dominio
 {
     public class Importadas:Compras
     {
-        public static double TasaImportacion { get; set; }
+        public static int TasaImportacion { get; set; }
         public bool EsDelSur { get; set; }
-        public double TasaDescuento { get; set; }
+        public int TasaDescuento { get; set; }
         public static string DescripcionSanitaria { get; set; }
     }
 }
